refactor: split parser out from FinalStepJsonProtoHaskellCabalLibraries
