+++ detectable/src/main/java/com/synopsys/integration/detectable/detectables/bazel/pipeline/Pipelines.java
@@ -32,48 +32,51 @@ import com.synopsys.integration.detectable.detectables.bazel.pipeline.step.Bazel
 import com.synopsys.integration.detectable.detectables.bazel.pipeline.step.BazelVariableSubstitutor;
 import com.synopsys.integration.detectable.detectables.bazel.pipeline.step.FinalStepColonSeparatedGavs;
 import com.synopsys.integration.detectable.detectables.bazel.pipeline.step.FinalStepJsonProtoHaskellCabalLibraries;
-import com.synopsys.integration.detectable.detectables.bazel.pipeline.step.IntermediateStepReplaceInEach;
+import com.synopsys.integration.detectable.detectables.bazel.pipeline.step.HaskellCabalLibraryJsonProtoParser;
 import com.synopsys.integration.detectable.detectables.bazel.pipeline.step.IntermediateStepExecuteBazelOnEach;
 import com.synopsys.integration.detectable.detectables.bazel.pipeline.step.IntermediateStepFilter;
 import com.synopsys.integration.detectable.detectables.bazel.pipeline.step.IntermediateStepParseEachXml;
+import com.synopsys.integration.detectable.detectables.bazel.pipeline.step.IntermediateStepReplaceInEach;
 import com.synopsys.integration.detectable.detectables.bazel.pipeline.step.IntermediateStepSplitEach;
 import com.synopsys.integration.exception.IntegrationException;
 
 public class Pipelines {
     private final Map<WorkspaceRule, Pipeline> availablePipelines = new HashMap<>();
 
-    public Pipelines(final BazelCommandExecutor bazelCommandExecutor, final BazelVariableSubstitutor bazelVariableSubstitutor,
-        final ExternalIdFactory externalIdFactory) {
-        final Pipeline mavenJarPipeline = (new PipelineBuilder())
-            .addIntermediateStep(new IntermediateStepExecuteBazelOnEach(bazelCommandExecutor, bazelVariableSubstitutor, Arrays.asList("query", "filter('@.*:jar', deps(${detect.bazel.target}))")))
-            .addIntermediateStep(new IntermediateStepSplitEach("\\s+"))
-            .addIntermediateStep(new IntermediateStepReplaceInEach("^@", ""))
-            .addIntermediateStep(new IntermediateStepReplaceInEach("//.*", ""))
-            .addIntermediateStep(new IntermediateStepReplaceInEach("^", "//external:"))
-            .addIntermediateStep(new IntermediateStepExecuteBazelOnEach(bazelCommandExecutor, bazelVariableSubstitutor, Arrays.asList("query", "kind(maven_jar, ${input.item})", "--output", "xml")))
-            .addIntermediateStep(new IntermediateStepParseEachXml("/query/rule[@class='maven_jar']/string[@name='artifact']", "value"))
-            .setFinalStep(new FinalStepColonSeparatedGavs(externalIdFactory))
-            .build();
+    public Pipelines(BazelCommandExecutor bazelCommandExecutor, BazelVariableSubstitutor bazelVariableSubstitutor,
+        ExternalIdFactory externalIdFactory) {
+        Pipeline mavenJarPipeline = (new PipelineBuilder())
+                                              .addIntermediateStep(new IntermediateStepExecuteBazelOnEach(bazelCommandExecutor, bazelVariableSubstitutor, Arrays.asList("query", "filter('@.*:jar', deps(${detect.bazel.target}))")))
+                                              .addIntermediateStep(new IntermediateStepSplitEach("\\s+"))
+                                              .addIntermediateStep(new IntermediateStepReplaceInEach("^@", ""))
+                                              .addIntermediateStep(new IntermediateStepReplaceInEach("//.*", ""))
+                                              .addIntermediateStep(new IntermediateStepReplaceInEach("^", "//external:"))
+                                              .addIntermediateStep(new IntermediateStepExecuteBazelOnEach(bazelCommandExecutor, bazelVariableSubstitutor, Arrays.asList("query", "kind(maven_jar, ${input.item})", "--output", "xml")))
+                                              .addIntermediateStep(new IntermediateStepParseEachXml("/query/rule[@class='maven_jar']/string[@name='artifact']", "value"))
+                                              .setFinalStep(new FinalStepColonSeparatedGavs(externalIdFactory))
+                                              .build();
         availablePipelines.put(WorkspaceRule.MAVEN_JAR, mavenJarPipeline);
 
-        final Pipeline mavenInstallPipeline = (new PipelineBuilder())
-            .addIntermediateStep(new IntermediateStepExecuteBazelOnEach(bazelCommandExecutor, bazelVariableSubstitutor, Arrays.asList("cquery", "--noimplicit_deps", "${detect.bazel.cquery.options}", "kind(j.*import, deps(${detect.bazel.target}))", "--output", "build")))
-            .addIntermediateStep(new IntermediateStepSplitEach("\n"))
-            .addIntermediateStep(new IntermediateStepFilter(".*maven_coordinates=.*"))
-            .addIntermediateStep(new IntermediateStepReplaceInEach(".*\"maven_coordinates=", ""))
-            .addIntermediateStep(new IntermediateStepReplaceInEach("\".*", ""))
-              .setFinalStep(new FinalStepColonSeparatedGavs(externalIdFactory))
-              .build();
+        Pipeline mavenInstallPipeline = (new PipelineBuilder())
+                                                  .addIntermediateStep(new IntermediateStepExecuteBazelOnEach(bazelCommandExecutor, bazelVariableSubstitutor,
+                                                      Arrays.asList("cquery", "--noimplicit_deps", "${detect.bazel.cquery.options}", "kind(j.*import, deps(${detect.bazel.target}))", "--output", "build")))
+                                                  .addIntermediateStep(new IntermediateStepSplitEach("\n"))
+                                                  .addIntermediateStep(new IntermediateStepFilter(".*maven_coordinates=.*"))
+                                                  .addIntermediateStep(new IntermediateStepReplaceInEach(".*\"maven_coordinates=", ""))
+                                                  .addIntermediateStep(new IntermediateStepReplaceInEach("\".*", ""))
+                                                  .setFinalStep(new FinalStepColonSeparatedGavs(externalIdFactory))
+                                                  .build();
         availablePipelines.put(WorkspaceRule.MAVEN_INSTALL, mavenInstallPipeline);
 
-        final Pipeline haskellCabalLibraryPipeline = (new PipelineBuilder())
-              .addIntermediateStep(new IntermediateStepExecuteBazelOnEach(bazelCommandExecutor, bazelVariableSubstitutor, Arrays.asList("cquery", "--noimplicit_deps", "${detect.bazel.cquery.options}", "kind(haskell_cabal_library, deps(${detect.bazel.target}))", "--output", "jsonproto")))
-              .setFinalStep(new FinalStepJsonProtoHaskellCabalLibraries())
-              .build();
+        Pipeline haskellCabalLibraryPipeline = (new PipelineBuilder())
+                                                         .addIntermediateStep(new IntermediateStepExecuteBazelOnEach(bazelCommandExecutor, bazelVariableSubstitutor,
+                                                             Arrays.asList("cquery", "--noimplicit_deps", "${detect.bazel.cquery.options}", "kind(haskell_cabal_library, deps(${detect.bazel.target}))", "--output", "jsonproto")))
+                                                         .setFinalStep(new FinalStepJsonProtoHaskellCabalLibraries(new HaskellCabalLibraryJsonProtoParser()))
+                                                         .build();
         availablePipelines.put(WorkspaceRule.HASKELL_CABAL_LIBRARY, haskellCabalLibraryPipeline);
     }
 
-    public Pipeline get(final WorkspaceRule bazelDependencyType) throws IntegrationException {
+    public Pipeline get(WorkspaceRule bazelDependencyType) throws IntegrationException {
         if (!availablePipelines.containsKey(bazelDependencyType)) {
             throw new IntegrationException(String.format("No pipeline found for dependency type %s", bazelDependencyType.getName()));
         }
